requirements/dev.txt
-r base.txt

flake8
pytest
pytest-cov
requests-mock
coverage
coverage-badge
isort
mypy
mypy-extensions
wheel
twine
codecov
